requirements.txt
streamlit==1.27.0
ai71==0.1.0
python-dotenv